requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
beautifulsoup4==4.12.2
Django==4.2.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Pillow==10.0.1
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-decouple==3.8
pytz==2023.3.post1
redis==5.0.1
soupsieve==2.5
sqlparse==0.4.4
